Rename Components::Image to InteractiveImage with real inheritance
Collapses the Components::Image hand-off + Components::Image::Interactive
pair into one class, Components::InteractiveImage < Components::Image::Base.
The hand-off existed only to give the top-level class Kit syntax while
avoiding a Zeitwerk self-reference trap (Components::Image can't inherit
from Components::Image::Base because the class being defined shares its
name with the Image namespace segment in its own superclass path).
InteractiveImage doesn't collide with that namespace, so it inherits
directly with no delegation needed.

Sweeps all 18 Image(...) Kit-syntax call sites to InteractiveImage(...),
Image#broadcast_interactive_sizes's direct .new call, and merges
image_test.rb + image/interactive_test.rb into interactive_image_test.rb.
