+++ app/controllers/observations/images_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Clicking on an image currently fires a GET to these actions... because it
-# comes from a link made by Components::Image::Interactive(link: url_args)
+# comes from a link made by Components::InteractiveImage(link: url_args)
 # with CRUD refactor, change component link to fire a POST somehow?
 
 module Observations
