+++ app/views/controllers/observations/namings/suggestions/show.rb
@@ -142,7 +142,7 @@ module Views::Controllers::Observations::Namings::Suggestions
     def render_suggestion_image(sugg)
       return if sugg.image_obs.blank?
 
-      Image(
+      InteractiveImage(
         user: @user,
         image: sugg.image_obs.thumb_image,
         image_link: image_path(id: sugg.image_obs.id)
